requirements.txt
# Core packages
numpy>=1.24,<2
ipython>=8.10
notebook
jupyterlab>=3.0
pandas==2.1.4
matplotlib
seaborn
scipy
plotly

# Kedro
kedro~=0.19.5
kedro-telemetry>=0.3.1
kedro-datasets
kedro-datasets[pickle]
kedro-datasets[pandas]
kedro-datasets[json]
kedro-datasets[yaml]
kedro-mlflow
kedro-viz

# Statistical Modeling
statsmodels==0.14.0
phik==0.12.3
shap
nannyml
evidently==0.6.6

# Validation & Profiling
great-expectations==0.17.23
ydata-profiling==4.16.1
visions>=0.7.4

# Geospatial
# folium
# for MarkerCluster plugin
# folium.plugins

# Machine Learning
scikit-learn
xgboost
optuna~=3.6.0

# Testing
pytest~=7.2
pytest-cov~=3.0
pytest-mock>=1.7.1,<2.0

# Linting & formatting
ruff~=0.1.8

# Hopsworks SDK
hopsworks[python]